move TemplateAnnotationRector to Sensio\FrameworkExtraBundle
