+++ src/config/level/sensio/framework-extra-bundle-30.yml
@@ -1,0 +1,2 @@
+rectors:
+    Rector\Rector\Contrib\Sensio\FrameworkExtraBundle\TemplateAnnotationRector: ~
