+++ tests/Rector/Contrib/Sensio/FrameworkExtraBundle/TemplateAnnotationRector/Correct/correct2.php.inc
