+++ tests/Rector/Contrib/Sensio/FrameworkExtraBundle/TemplateAnnotationRector/Correct/correct6.php.inc
