+++ tests/Rector/Contrib/Sensio/FrameworkExtraBundle/TemplateAnnotationRector/Source/rector.yml
@@ -1,0 +1,2 @@
+rectors:
+    Rector\Rector\Contrib\Sensio\FrameworkExtraBundle\TemplateAnnotationRector: ~
