+++ tests/Rector/Contrib/Sensio/FrameworkExtraBundle/TemplateAnnotationRector/Wrong/wrong6.php.inc
