+++ tests/Rector/Contrib/Sensio/FrameworkExtraBundle/TemplateAnnotationRector/Wrong/wrong8.php.inc
