--- tests/Rector/Contrib/Symfony/HttpKernel/TemplateAnnotationRector/Source/rector.yml
@@ -1,2 +1,0 @@
-rectors:
-    Rector\Rector\Contrib\Symfony\HttpKernel\TemplateAnnotationRector: ~
